+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Setting to change note visibility, Show in taskbar, hide in taskbar or hide in taskbar and task switcher (Alt+Tab and Win+Tab).
 - Database migration system, seamlessly upgrade the database after upgrading the application.
 - Text Lock, toggle Locked from the text box context menu to prevent note text being edited.
+- URL tool, encode and decode text for use in URls.
 - 
 
 ### Changed